+++ src/main/java/org/braekpo1nt/mctmanager/games/game/spleef/config/SpleefStorageUtil.java
@@ -9,7 +9,7 @@ import net.kyori.adventure.text.serializer.gson.GsonComponentSerializer;
 import org.braekpo1nt.mctmanager.Main;
 import org.braekpo1nt.mctmanager.games.game.config.GameConfigStorageUtil;
 import org.braekpo1nt.mctmanager.games.game.spleef.DecayStage;
-import org.braekpo1nt.mctmanager.games.game.spleef.Powerup;
+import org.braekpo1nt.mctmanager.games.game.spleef.powerup.Powerup;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
